fix: Set overflow property when height auto.
The overflow property by default is set to `hidden`, because when we
start with a default container we have a height set at `0px`. If the
overflow property wasn't set to `hidden` then we would see the content.

When animating to a height of `auto` it makes sense to allow any
overflowed elements to be visible once the animation has ended. Things
such as absolutely positioned menus etc can now be visible without being
clipped by the container.