fix: FixedAtAuto used with heightAt
Add animate property

The animate property is set to False by the heightAt function which
bypasses animation events and sets the height value immediately.

Set values for heightAt function

CALCULATEDHEIGHT: Stays the same if the target is FixedAtAuto.

FORCE: We need to force an animation frame as calculating the
FixedAtAuto values requires a Dom query.

ANIMATE: Since we are in heightAt we don't want to cause animations.

Set duration

When we don't want animation events we set the duration to 0ms.

This means no transition events will fire.

Add simple test for FixedAtAuto